requirements.txt
asgiref==3.11.1
Django==5.2.11
django-rest-framework==0.1.0
djangorestframework==3.16.1
sqlparse==0.5.5
tzdata==2025.3
